Revamp iOS dylib injector payload so it behaves properly.
Previous attempt at doing pthread setup turned out to be a
flawed solution and dead end, resulting in undefined behavior
whenever pthread decided to reap thread contexts.

So now we revert back to something close to the original approach, where the naked mach-thread's pthread state is
just a zero-filled structure, and it just calls pthread_create
to spawn a proper pthread, where dlopen() and friends get called. This time around we also put a guard page at the
bottom of the stack, so stack overflows get caught as early
as possible.